Add check_config helper (#24557)
* check_config

* no ignore

* tests

* try tests again
