+++ tests/common.py
@@ -16,7 +16,6 @@ from unittest.mock import MagicMock, Mock, patch
 
 import homeassistant.util.dt as date_util
 import homeassistant.util.yaml.loader as yaml_loader
-import homeassistant.util.yaml.dumper as yaml_dumper
 
 from homeassistant import auth, config_entries, core as ha, loader
 from homeassistant.auth import (
@@ -682,7 +681,6 @@ def patch_yaml_files(files_dict, endswith=True):
         raise FileNotFoundError("File not found: {}".format(fname))
 
     return patch.object(yaml_loader, 'open', mock_open_f, create=True)
-    return patch.object(yaml_dumper, 'open', mock_open_f, create=True)
 
 
 def mock_coro(return_value=None, exception=None):
